Ouroboros.Consensus.Network.NodeToClient

Description

Intended for qualified import

Handlers

data Handlers m peer blk

Protocol handlers for node-to-client (local) communication

Constructors

Handlers

Fields

hChainSyncServer :: Follower m blk (WithPoint blk (Serialised blk)) -> ChainSyncServer (Serialised blk) (Point blk) (Tip blk) m ()

hTxSubmissionServer :: LocalTxSubmissionServer (GenTx blk) (ApplyTxErr blk) m ()

hStateQueryServer :: LocalStateQueryServer blk (Point blk) (Query blk) m ()

hTxMonitorServer :: LocalTxMonitorServer (GenTxId blk) (GenTx blk) SlotNo m ()

mkHandlers :: forall m blk addrNTN addrNTC. (IOLike m, LedgerSupportsMempool blk, LedgerSupportsProtocol blk, QueryLedger blk, ConfigSupportsNode blk) => NodeKernelArgs m addrNTN addrNTC blk -> NodeKernel m addrNTN addrNTC blk -> Handlers m addrNTC blk

Codecs

type ClientCodecs blk m = Codecs' blk blk DeserialiseFailure m ByteString ByteString ByteString ByteString

type Codecs blk e m bCS bTX bSQ bTM = Codecs' blk (Serialised blk) e m bCS bTX bSQ bTM

data Codecs' blk serialisedBlk e m bCS bTX bSQ bTM

Node-to-client protocol codecs needed to run Handlers.

Constructors

Codecs

Fields

cChainSyncCodec :: Codec (ChainSync serialisedBlk (Point blk) (Tip blk)) e m bCS

cTxSubmissionCodec :: Codec (LocalTxSubmission (GenTx blk) (ApplyTxErr blk)) e m bTX

cStateQueryCodec :: Codec (LocalStateQuery blk (Point blk) (Query blk)) e m bSQ

cTxMonitorCodec :: Codec (LocalTxMonitor (GenTxId blk) (GenTx blk) SlotNo) e m bTM

type DefaultCodecs blk m = Codecs' blk (Serialised blk) DeserialiseFailure m ByteString ByteString ByteString ByteString

clientCodecs :: forall m blk. (MonadST m, SerialiseNodeToClientConstraints blk, ShowQuery (BlockQuery blk), StandardHash blk, Serialise (HeaderHash blk)) => CodecConfig blk -> BlockNodeToClientVersion blk -> NodeToClientVersion -> ClientCodecs blk m

Protocol codecs for the node-to-client protocols which serialise deserialise blocks in chain-sync/ protocol.

defaultCodecs :: forall m blk. (MonadST m, SerialiseNodeToClientConstraints blk, ShowQuery (BlockQuery blk), StandardHash blk, Serialise (HeaderHash blk)) => CodecConfig blk -> BlockNodeToClientVersion blk -> NodeToClientVersion -> DefaultCodecs blk m

Protocol codecs for the node-to-client protocols

We pass the BlockConfig here, even though it is currently unused. If at any point we want to introduce local protocols that for example send Byron blocks or headers across, we will need to have the epoch size, which comes from the Byron config. Unlike the full TopLevelConfig, it should not be difficult for a wallet to construct the BlockConfig.

NOTE: Somewhat confusingly, pcChainSyncCodec currently does send Byron blocks across, but it does not deserialize them (the user of the codec is itself responsible for doing that), which is why it currently does not need the config.

Implementation mode: currently none of the consensus encoders/decoders do anything different based on the version, so _version is unused; it's just that not all codecs are used, depending on the version number.

identityCodecs :: (Monad m, QueryLedger blk) => Codecs blk CodecFailure m (AnyMessage (ChainSync (Serialised blk) (Point blk) (Tip blk))) (AnyMessage (LocalTxSubmission (GenTx blk) (ApplyTxErr blk))) (AnyMessage (LocalStateQuery blk (Point blk) (Query blk))) (AnyMessage (LocalTxMonitor (GenTxId blk) (GenTx blk) SlotNo))

Identity codecs used in tests.

ClientCodecs

Tracers

type Tracers m peer blk e = Tracers' peer blk e (Tracer m)

A record of Tracers for the different protocols.

data Tracers' peer blk e f

Constructors

Tracers

Fields

tChainSyncTracer :: f (TraceLabelPeer peer (TraceSendRecv (ChainSync (Serialised blk) (Point blk) (Tip blk))))

tTxSubmissionTracer :: f (TraceLabelPeer peer (TraceSendRecv (LocalTxSubmission (GenTx blk) (ApplyTxErr blk))))

tStateQueryTracer :: f (TraceLabelPeer peer (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk))))

tTxMonitorTracer :: f (TraceLabelPeer peer (TraceSendRecv (LocalTxMonitor (GenTxId blk) (GenTx blk) SlotNo)))

nullTracers :: Monad m => Tracers m peer blk e

Use a nullTracer for each protocol.

showTracers :: (Show peer, Show (GenTx blk), Show (GenTxId blk), Show (ApplyTxErr blk), ShowQuery (BlockQuery blk), HasHeader blk) => Tracer m String -> Tracers m peer blk e

Applications

type App m peer bytes a = peer -> Channel m bytes -> m (a, Maybe bytes)

A node-to-client application

data Apps m peer bCS bTX bSQ bTM a

Applications for the node-to-client (i.e., local) protocols

See MuxApplication

Constructors

Apps

Fields

aChainSyncServer :: App m peer bCS a

Start a local chain sync server.

aTxSubmissionServer :: App m peer bTX a

Start a local transaction submission server.

aStateQueryServer :: App m peer bSQ a

Start a local state query server.

aTxMonitorServer :: App m peer bTM a

Start a local transaction monitor server

mkApps :: forall m addrNTN addrNTC blk e bCS bTX bSQ bTM. (IOLike m, Exception e, ShowProxy blk, ShowProxy (ApplyTxErr blk), ShowProxy (BlockQuery blk), ShowProxy (GenTx blk), ShowProxy (GenTxId blk), ShowQuery (BlockQuery blk)) => NodeKernel m addrNTN addrNTC blk -> Tracers m addrNTC blk e -> Codecs blk e m bCS bTX bSQ bTM -> Handlers m addrNTC blk -> Apps m addrNTC bCS bTX bSQ bTM ()

Construct the NetworkApplication for the node-to-client protocols

Projections

responder :: NodeToClientVersion -> Apps m (ConnectionId peer) b b b b a -> OuroborosApplicationWithMinimalCtx 'ResponderMode peer b m Void a

A projection from NetworkApplication to a server-side OuroborosApplication for the node-to-client protocols.